requirements.txt
flask==2.3.3
flask-cors==4.0.0
opencv-python==4.8.1.78
ultralytics==8.0.196
pillow==10.0.1
numpy==1.24.3
werkzeug==2.3.7
python-dotenv==1.0.0
gunicorn==21.2.0
psutil==5.9.5
requests==2.31.0
